compressing_classifiers_and_MLPs/installingImageNet/requirements.txt
datasets
pillow
huggingface_hub